pyproject.toml
[project]
name = "whisper_servers"
version = "0.1.0"
description = "Dual MLX Whisper Servers for real-time and batch transcription"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.105.0",
    "uvicorn>=0.24.0", 
    "starlette>=0.27.0",
    "mlx-whisper>=0.4.2",
    "mlx-audio>=0.2.1",
    "rich>=13.6.0",
    "textual>=0.42.0",
    "ffmpeg-python>=0.2.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.2",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "mypy>=1.6.1",
]

[project.scripts]
whisper-server = "whisper_servers.__main__:main"
whisper-batch-server = "whisper_servers.scripts.run_batch:main"
whisper-realtime-server = "whisper_servers.scripts.run_realtime:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
